requirements.txt
flask==3.0.3
langchain==0.3.7
langchain-community==0.3.7
langchain-groq==0.2.1
python-dotenv==1.0.1
pypdf==5.1.0
markdown==3.7
pytest==8.3.3
gunicorn==23.0.0
